HotLogger & LoggerBase: new feature, echo to console: Allow you to specify that all logging is echoed to the console. In some cases using the built in feature of capturing logging using logging library results in poor scannable output. That's still available, this is just a very simple alternative that gives you good log output when using test runner like Resharper.
traceTitle: 
Quick method that writes newline, title, and underlines the title to the logs. Improves scannability. 

GetLogger now checks if logging enabled, if not returns null logger. 

StringExtensions:
new method ( UnderlineText) : used for pretty logging, improves scannability of logfiles. 

SeperateFileProcessorPerEndpoint:
Commented out for now, started working on this, then had to pause this to complete other changes first in order to be able to achieve this.