Endpoint request matching (mechanism for determining successful request match that needs to be processed) in custom HttpModules is now fully unit testable. See EndpointRequestMatcherTests.cs
Added ability to have runtime contract checking of nulls. DebugContract.cs
For Tests and sample usage see DebugContractTests.cs
This is slow because it uses reflection so is marked with Conditional("DEBUG") attribute.

Added QuickAssert which makes the unit tests much more scannable and readable. Was using Fluent Assertions, but the syntax was cluttered for checking a complex object with many properties. See ModuleConfigurationBaseTests.cs for example of usage.